# Packed MTF mask