--- a/megatron/model/utils.py
+++ b/megatron/model/utils.py
@@ -40,7 +40,7 @@ def scaled_init_method_normal(sigma, num_layers):
 
 
 def attention_mask_func(attention_scores, attention_mask):
-    attention_scores.masked_fill_(attention_mask, -10000.0)
+    attention_scores.masked_fill_(attention_mask, torch.finfo(attention_scores.dtype).min)
     return attention_scores